fix di

+++ app/src/main/java/com/vaskorr/weatherapp/data/json/Location.kt
@@ -9,7 +9,7 @@ data class Location (
   @SerializedName("region"          ) var region         : String? = null,
   @SerializedName("country"         ) var country        : String? = null,
   @SerializedName("lat"             ) var lat            : Double? = null,
-  @SerializedName("lon"             ) var lon            : Int?    = null,
+  @SerializedName("lon"             ) var lon            : Double? = null,
   @SerializedName("tz_id"           ) var tzId           : String? = null,
   @SerializedName("localtime_epoch" ) var localtimeEpoch : Int?    = null,
   @SerializedName("localtime"       ) var localtime      : String? = null
